chore: Rename skate() to define() for named exports. You can still import skate from skatejs as a default.
