+++ src/index.js
@@ -2,6 +2,7 @@ import assign from 'object-assign';
 
 // Public API
 import create from './api/create';
+import define from './api/define';
 import emit from './api/emit';
 import factory from './api/factory';
 import fragment from './api/fragment';
@@ -9,7 +10,6 @@ import init from './api/init';
 import link from './api/link';
 import prop, * as props from './api/prop';
 import ready from './api/ready';
-import skate from './api/skate';
 import state from './api/state';
 import * as symbols from './api/symbols';
 import vdom, * as vdomElements from './api/vdom';
@@ -18,9 +18,10 @@ import version from './api/version';
 assign(prop, props);
 assign(vdom, vdomElements);
 
-export default skate;
+export default define;
 export {
   create,
+  define,
   emit,
   factory,
   fragment,
@@ -28,7 +29,6 @@ export {
   link,
   prop,
   ready,
-  skate,
   state,
   symbols,
   vdom,
